pyproject.toml
[project]
name = "vesper"
version = "0.1.1"
description = "The CUDA accelerated version of VESPER"
readme = "README.md"
authors = [
    { name = "Han Zhu", email = "zhu773@purdue.edu" }
]
requires-python = ">=3.11"
dependencies = [
    "biopython>=1.85",
    "gemmi>=0.7.1",
    "mrcfile>=1.5.4",
    "numba>=0.61.2",
    "numba-progress>=0.0.10",
    "numpy",
    "pydssp>=0.9.1",
    "pyfftw>=0.15.1",
    "scikit-learn",
    "scipy>=1.16.3",
    "torch>=2",
    "torchvision",
    "tqdm>=4.67.1",
    "typer>=0.20.0",
]

[project.scripts]
vesper = "vesper.cli:app"

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", marker = "sys_platform != 'linux'" },
  { index = "pytorch-cu126", marker = "sys_platform == 'linux'" },
]
torchvision = [
  { index = "pytorch-cpu", marker = "sys_platform != 'linux'" },
  { index = "pytorch-cu126", marker = "sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[dependency-groups]
dev = [
    "prek>=0.2.22",
    "ruff>=0.14.9",
    "ty>=0.0.3",
]
